+++ railties/lib/rails/commands/console/console_command.rb
@@ -70,9 +70,6 @@ module Rails
       class_option :sandbox, aliases: "-s", type: :boolean, default: false,
         desc: "Rollback database modifications on exit."
 
-      class_option :environment, aliases: "-e", type: :string,
-        desc: "Specifies the environment to run this console under (test/development/production)."
-
       def initialize(args = [], local_options = {}, config = {})
         console_options = []