+++ railties/CHANGELOG.md
@@ -1,3 +1,29 @@
+*   Properly expand shortcuts for environment's name running the `console`
+    and `dbconsole` commands.
+
+    *Robin Dupret*
+
+*   Passing the environment's name as a regular argument to the
+    `rails dbconsole` and `rails console` commands is deprecated.
+    The `-e` option should be used instead.
+
+    Previously:
+
+        $ bin/rails dbconsole production
+
+    Now:
+
+        $ bin/rails dbconsole -e production
+
+    *Robin Dupret*, *Kasper Timm Hansen*
+
+*   Allow to pass a custom connection name to the `rails dbconsole`
+    command when using a 3-level database configuration.
+
+        $ bin/rails dbconsole -c replica
+
+    *Robin Dupret*, *Jeremy Daer*
+
 *   Skip unused components when running `bin/rails app:update`.
 
     If the initial app generation skipped Action Cable, Active Record etc.,